\name{sitools-package}
\alias{sitools-package}
\alias{sitools}
\docType{package}
\title{SI tools}
\description{tools to handle prefixes as defined in Systeme International d Unites.}

\author{
  Jonas Stein, thanks to Ben Tupper, Roland Fu(sz), Ryan C. Thompson
Maintainer: Jonas Stein <news@jonasstein.de>
}
\references{\url{http://www.bipm.org/en/si/}}
\keyword{ package }
\examples{
f2si(0.001)
}
