\name{sitools-package}
\alias{sitools-package}
\alias{sitools}
\docType{package}
\title{SI tools}
\description{tools to handle prefixes as defined in Systeme International d Unites.}
\details{
\tabular{ll}{
Package: \tab sitools\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2012-01-23\cr
License: \tab GPL-3 \cr
LazyLoad: \tab yes\cr
}

}
\author{
  Jonas Stein, thanks to Ben Tupper, Roland Fu(sz)
Maintainer: Jonas Stein <news@jonasstein.de>
}
\references{\url{http://www.bipm.org/en/si/}}
\keyword{ package }
\examples{
f2si(0.001)
}
